+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Removed `federation.fep_c0e0_emoji_react_enabled` configuration parameter.
 - Removed `http_cors_allow_all` configuration parameter.
+- Removed support for payment links with `https://w3id.org/valueflows/Proposal` rel type.
 
 ## [4.20.0] - 2026-03-19